change user route to auth route, add new route and auth middleware
