+++ server/models/userModel.ts
@@ -8,31 +8,36 @@ interface User {
   avatar?: string;
 }
 
-const UserSchema = new Schema<User>({
-  email: {
-    type: String,
-    unique: true,
-    required: true,
-    min: MIN_LENGTH_6,
-    max: MAX_LENGTH_255,
+const UserSchema = new Schema<User>(
+  {
+    email: {
+      type: String,
+      unique: true,
+      required: true,
+      min: MIN_LENGTH_6,
+      max: MAX_LENGTH_255,
+    },
+    name: {
+      type: String,
+      required: true,
+      min: MIN_LENGTH_6,
+      max: MAX_LENGTH_255,
+    },
+    avatar: {
+      base64: String,
+      imageFormat: String,
+    },
+    password: {
+      type: String,
+      required: true,
+      min: MIN_LENGTH_6,
+      max: MAX_LENGTH_255,
+    },
   },
-  name: {
-    type: String,
-    required: true,
-    min: MIN_LENGTH_6,
-    max: MAX_LENGTH_255,
-  },
-  avatar: {
-    base64: String,
-    imageFormat: String,
-  },
-  password: {
-    type: String,
-    required: true,
-    min: MIN_LENGTH_6,
-    max: MAX_LENGTH_255,
-  },
-});
+  {
+    timestamps: true,
+  }
+);
 
 const UserModel = model<User>('User', UserSchema);
